autoria_scraper/requirements.txt
Django>=4.2,<5.0
django-environ>=0.10.0
psycopg2-binary>=2.9.5
requests>=2.28.0
beautifulsoup4>=4.11.1
celery>=5.3.0
redis>=4.5.0
python-dotenv>=1.0.0